+++ test/test_intersects.h
@@ -63,7 +63,7 @@ void load_data(void){
     */
 
     MemMergeT temp = allocMemMergeT(pt.len);
-    int inv = MERGE_SORT_COUNT_INVERSION(pt, 0, pt.len-1, temp);
+    ulong inv = MERGE_SORT_COUNT_INVERSION(pt, 0, pt.len-1, temp);
 
     TEST_CHECK(pt.p[0] == -14);
     TEST_CHECK(pt.l[0].val == -16);
@@ -106,11 +106,11 @@ void testData(void){
     TriProb tp = ParseData(filepath);
 
     //Solve Problem
-    int ans = get_intersects(tp.n, tp.ps, tp.qs, tp.rs);
+    ulong ans = get_intersects(tp.n, tp.ps, tp.qs, tp.rs);
 
     //Check Answer
     TEST_CHECK(ans>=0);
-    TEST_MSG("EST: %d ; REAL: %d", ans, tp.sol);
+    TEST_MSG("EST: %ld ; REAL: %d", ans, tp.sol);
     
     free(tp.ps);
     free(tp.qs);
